requirements.txt
mysql-connector-python 
pandas
yagmail 
python-dotenv